+++ presentation/src/main/java/com/threegap/bitnagil/presentation/mypage/MyPageViewModel.kt
@@ -2,8 +2,8 @@ package com.threegap.bitnagil.presentation.mypage
 
 import androidx.lifecycle.ViewModel
 import com.threegap.bitnagil.domain.user.usecase.FetchUserProfileUseCase
-import com.threegap.bitnagil.presentation.mypage.model.MyPageSideEffect
-import com.threegap.bitnagil.presentation.mypage.model.MyPageState
+import com.threegap.bitnagil.presentation.mypage.contract.MyPageSideEffect
+import com.threegap.bitnagil.presentation.mypage.contract.MyPageState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
@@ -14,7 +14,7 @@ import javax.inject.Inject
 class MyPageViewModel @Inject constructor(
     private val fetchUserProfileUseCase: FetchUserProfileUseCase,
 ) : ContainerHost<MyPageState, MyPageSideEffect>, ViewModel() {
-    override val container: Container<MyPageState, MyPageSideEffect> = container(initialState = MyPageState.Init)
+    override val container: Container<MyPageState, MyPageSideEffect> = container(initialState = MyPageState.INIT)
 
     init {
         loadMyPageInfo()
