+++ presentation/src/main/java/com/threegap/bitnagil/presentation/mypage/contract/MyPageSideEffect.kt
@@ -1,0 +1,3 @@
+package com.threegap.bitnagil.presentation.mypage.contract
+
+sealed interface MyPageSideEffect
