+++ presentation/src/main/java/com/threegap/bitnagil/presentation/mypage/contract/MyPageState.kt
@@ -1,0 +1,13 @@
+package com.threegap.bitnagil.presentation.mypage.contract
+
+data class MyPageState(
+    val name: String,
+    val profileUrl: String,
+) {
+    companion object {
+        val INIT = MyPageState(
+            name = "",
+            profileUrl = ""
+        )
+    }
+}
