--- presentation/src/main/java/com/threegap/bitnagil/presentation/mypage/model/MyPageSideEffect.kt
@@ -1,3 +1,0 @@
-package com.threegap.bitnagil.presentation.mypage.model
-
-sealed class MyPageSideEffect
