--- presentation/src/main/java/com/threegap/bitnagil/presentation/mypage/model/MyPageState.kt
@@ -1,10 +1,0 @@
-package com.threegap.bitnagil.presentation.mypage.model
-
-data class MyPageState(
-    val name: String,
-    val profileUrl: String,
-) {
-    companion object {
-        val Init = MyPageState(name = "", profileUrl = "")
-    }
-}
